+++ src/main/java/lapr/project/model/Administrator.java
@@ -8,9 +8,9 @@ public class Administrator {
     private int pharmacyId;
     private String name;
     private int nif;
-    private int socialSecurity;
+    private long socialSecurity;
 
-    public Administrator(String email, int pharmacyId, String name, int nif, int socialSecurity) {
+    public Administrator(String email, int pharmacyId, String name, int nif, long socialSecurity) {
 
         this.email = email;
         this.pharmacyId = pharmacyId;
@@ -43,7 +43,7 @@ public class Administrator {
         return nif;
     }
 
-    public int getSocialSecurity() {
+    public long getSocialSecurity() {
 
         return socialSecurity;
     }
@@ -68,7 +68,7 @@ public class Administrator {
         this.nif = nif;
     }
 
-    public void setSocialSecurity(int socialSecurity) {
+    public void setSocialSecurity(long socialSecurity) {
 
         this.socialSecurity = socialSecurity;
     }
